% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_weights_function.R
\name{generate_rep_weights}
\alias{generate_rep_weights}
\title{My Function}
\usage{
generate_rep_weights(
  data,
  method = "jackknife",
  num_reps = 50,
  strata_var = NULL
)
}
\arguments{
\item{data}{Dataframe containing survey data.}

\item{method}{Method used to generate replicate weights: jackknife, bootstrap, and brr.}

\item{num_reps}{Number of replications used to generate replicate weights.}

\item{strata_var}{Variable used for stratification.}
}
\value{
The list of generated replicate weights.
}
\description{
This function generates replicate weights for complex survey data.
}
\examples{
data <- data.frame(
  id = 1:10,
  weight = runif(10, 1, 5),
  gender = rep(c("Male", "Female"), 5),
  age = sample(20:50, 10, replace = TRUE),
  cleaned_race_eth = sample(c("White", "Black", "Asian", "Other"), 10, replace = TRUE)
)
generate_rep_weights(data, method = "jackknife")
generate_rep_weights(data, method = "bootstrap", num_reps = 50)
generate_rep_weights(data, method = "brr", strata_var = "cleaned_race_eth")

}
